Improve glimmer queries  (#6)
* Improve glimmer highlight queries

The previous queries I suspect was based on neovim’s queries and used
captures not supported by Zed. Zed uses the latest matching query, which
I've leverage to make it easier to follow, removing all not-eq style
predicates.

* Implement highlighting brackets queries

This highlights the opening/closing brackets when the cursor is on one
of the pair.

* Update indentation queries

The existing indentation didn't capture the way Zed expected.